feat: solve sensory feedback place coding!
YESS!
Very exited to refactor some of this code tomorrow!
I will use named tuples in step, add docstrings, decide how to use screen_height and screen_width in the main script.
But man o man, I'm this close to conclude my part and get to the juice of it working on the model!